pyproject.toml
[project]
name = "server-status-cv-py"
version = "0.1.0"
description = "Server status client"
authors = [
    {name = "snowykami", email = "snowykami@outlook.com"},
]
dependencies = [
    "requests>=2.32.3",
    "psutil>=6.0.0",
    "arclet-alconna>=1.8.30",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
